+++ test/unittest/src/performance_test.cpp
@@ -29,7 +29,7 @@ PerformanceTest::~PerformanceTest()
 {
 }
 
-long long PerformanceTest::GetCurrentTime()
+int64_t PerformanceTest::GetCurrentTime()
 {
     int subtle = 1000000;
     struct timeval time;
@@ -37,7 +37,7 @@ long long PerformanceTest::GetCurrentTime()
     return (time.tv_sec * subtle + time.tv_usec);
 }
 
-int PerformanceTest::CalcTime(long long startTime, long long endTime)
+int PerformanceTest::CalcTime(int64_t startTime, int64_t endTime)
 {
     return (int)(endTime - startTime);
 }
@@ -95,7 +95,7 @@ HWTEST_F(PerformanceTest, raw_contact_insert_performance_test_900, testing::ext:
         values.push_back(rawContactValues);
     }
     HILOG_INFO("raw_contact_insert_performance_test_900  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int batchInsertCode = contactsDataAbility.BatchInsert(uriRawContact, values);
@@ -127,7 +127,7 @@ HWTEST_F(PerformanceTest, raw_contact_update_performance_test_1000, testing::ext
     predicates.And();
     predicates.EqualTo("is_deleted", "0");
     HILOG_INFO("raw_contact_update_performance_test_1000  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int updateCode = contactsDataAbility.Update(uriRawContact, updateValues, predicates);
@@ -160,7 +160,7 @@ HWTEST_F(PerformanceTest, raw_contact_query_performance_test_1100, testing::ext:
     predicates.And();
     predicates.EqualTo("is_deleted", "0");
     HILOG_INFO("raw_contact_query_performance_test_1100  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     std::shared_ptr<OHOS::NativeRdb::AbsSharedResultSet> resultSet =
@@ -192,7 +192,7 @@ HWTEST_F(PerformanceTest, raw_contact_delete_performance_test_1200, testing::ext
     predicates.And();
     predicates.EqualTo("is_deleted", "0");
     HILOG_INFO("raw_contact_delete_performance_test_1200  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int deleteCode = contactsDataAbility.Delete(uriRawContact, predicates);
@@ -239,7 +239,7 @@ HWTEST_F(PerformanceTest, contact_data_insert_performance_test_1300, testing::ex
         values.push_back(contactDataValues);
     }
     HILOG_INFO("contact_data_insert_performance_test_1300  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int batchInsertCode = contactsDataAbility.BatchInsert(uriContactData, values);
@@ -266,7 +266,7 @@ HWTEST_F(PerformanceTest, contact_data_update_performance_test_1400, testing::ex
     updateValues.PutString("syn_1", "test");
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int updateCode = contactsDataAbility.Update(uriContactData, updateValues, predicates);
@@ -294,7 +294,7 @@ HWTEST_F(PerformanceTest, contact_data_query_performance_test_1500, testing::ext
     columns.push_back("detail_info");
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     std::shared_ptr<OHOS::NativeRdb::AbsSharedResultSet> resultSet =
@@ -323,7 +323,7 @@ HWTEST_F(PerformanceTest, contact_data_delete_performance_test_1600, testing::ex
     OHOS::Uri uriContactData(ContactsUri::CONTACT_DATA);
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int deleteCode = contactsDataAbility.Delete(uriContactData, predicates);
@@ -356,7 +356,7 @@ HWTEST_F(PerformanceTest, calllog_insert_performance_test_100, testing::ext::Tes
         calllogValues.PutString("phone_number", std::to_string(i + 1));
         values.push_back(calllogValues);
     }
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int batchInsertCode = calllogAbility.BatchInsert(uriCalllog, values);
@@ -383,7 +383,7 @@ HWTEST_F(PerformanceTest, calllog_update_performance_test_200, testing::ext::Tes
     updateValues.PutInt("answer_state", 1);
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int updateCode = calllogAbility.Update(uriCalllog, updateValues, predicates);
@@ -411,7 +411,7 @@ HWTEST_F(PerformanceTest, calllog_query_performance_test_300, testing::ext::Test
     columns.push_back("phone_number");
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     std::shared_ptr<OHOS::NativeRdb::AbsSharedResultSet> resultSet =
@@ -440,7 +440,7 @@ HWTEST_F(PerformanceTest, calllog_delete_performance_test_400, testing::ext::Tes
     OHOS::Uri uriCalllog(CallLogUri::CALL_LOG);
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int deleteCode = calllogAbility.Delete(uriCalllog, predicates);
@@ -474,7 +474,7 @@ HWTEST_F(PerformanceTest, voicemail_insert_performance_test_500, testing::ext::T
         values.push_back(voicemailValues);
     }
     HILOG_INFO("voicemail_insert_performance_test_500  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int batchInsertCode = voicemailAbility.BatchInsert(uriVoiceMail, values);
@@ -501,7 +501,7 @@ HWTEST_F(PerformanceTest, voicemail_update_performance_test_600, testing::ext::T
     updateValues.PutString("origin_type", "origin");
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int updateCode = voicemailAbility.Update(uriVoiceMail, updateValues, predicates);
@@ -531,7 +531,7 @@ HWTEST_F(PerformanceTest, voicemail_query_performance_test_700, testing::ext::Te
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
     HILOG_INFO("voicemail_query_performance_test_700  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     std::shared_ptr<OHOS::NativeRdb::AbsSharedResultSet> resultSet =
@@ -561,7 +561,7 @@ HWTEST_F(PerformanceTest, voicemail_delete_performance_test_800, testing::ext::T
     OHOS::NativeRdb::DataAbilityPredicates predicates;
     predicates.GreaterThan("id", "0");
     HILOG_INFO("voicemail_delete_performance_test_800  start! ");
-    long long startTime, endTime;
+    int64_t startTime, endTime;
     int elaps;
     startTime = GetCurrentTime();
     int deleteCode = voicemailAbility.Delete(uriVoiceMail, predicates);
